Main program 2 (Smaller model)/requirements.txt
PyQt5==5.15.4
torch==2.0.0
opencv-python==4.7.0.72
mysql-connector-python==8.0.32
requests==2.30.0
imutils==0.5.4
Pillow==9.3.0
torchvision==0.15.1
torchaudio==2.0.1
QT-PyQt-PySide-Custom-Widgets==0.6.4
ipython==8.7.0
ultralytics==8.0.20
supervision==0.8.0
pyodbc==4.0.35
pyqt-toast==0.0.15
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pyqt5-tools==5.15.4.3.2
PySide2==5.15.2.1
PySide6-Essentials==6.4.3
rlPyCairo==0.2.0
pycairo==1.23.0
CairoSVG==2.7.0
pipwin==0.5.2
reportlab==4.0.0